Remove the state normalization
- Change the strucuture of modifying mdp_info
