+++ mushroom_rl/environments/pybullet_envs/air_hockey/planar/env_single.py
@@ -1,6 +1,7 @@
 import numpy as np
 import pybullet_utils.transformations as transformations
-
+from mushroom_rl.core import MDPInfo
+from mushroom_rl.utils.spaces import Box
 from mushroom_rl.environments.pybullet_envs.air_hockey.planar.env_base import AirHockeyPlanarBase, PyBulletObservationType
 
 
@@ -16,8 +17,16 @@ class AirHockeyPlanarSingle(AirHockeyPlanarBase):
 
         self._client.resetDebugVisualizerCamera(cameraDistance=1.5, cameraYaw=-90.0, cameraPitch=-45.0,
                                                 cameraTargetPosition=[-0.5, 0., 0.])
-        self.obs_norm_scale = np.concatenate([[1.0, 0.5, np.pi, 1., 1., np.pi * 2],
-                                              self.joints.limits()[1], self.joints.velocity_limits()])
+
+    def _modify_mdp_info(self, mdp_info):
+        obs_idx = [0, 1, 2, 7, 8, 9, 13, 14, 15, 16, 17, 18]
+        obs_low = mdp_info.observation_space.low[obs_idx]
+        obs_high = mdp_info.observation_space.high[obs_idx]
+        obs_low[0:3] = [-1, -0.5, -np.pi]
+        obs_high[0:3] = [1, 0.5, np.pi]
+        observation_space = Box(low=obs_low, high=obs_high)
+        return MDPInfo(observation_space, mdp_info.action_space, mdp_info.gamma, mdp_info.horizon)
+
 
     def _create_observation(self, state):
         puck_pose = self.get_sim_state(state, "puck", PyBulletObservationType.BODY_POS)
@@ -50,8 +59,6 @@ class AirHockeyPlanarSingle(AirHockeyPlanarBase):
             robot_vel = alpha * robot_vel + (1 - alpha) * self.obs_prev[9:12]
 
         self.obs_prev = np.concatenate([puck_pose_2d, puck_vel_2d, robot_pos, robot_vel])
-
-        self.obs_prev = self.obs_prev / self.obs_norm_scale
         return self.obs_prev
 
     def _puck_2d_in_robot_frame(self, puck_in, robot_frame, type='pose'):
